+++ CHANGES.rst
@@ -34,6 +34,9 @@ Changelog
 
 -   Disable sending alerts on production, leave them on for other instances.
 
+-   Add a form to the Flask console to create BAYESTAR sky maps with selected
+    detectors disabled.
+
 2.1.10 "Bunyip" (12-05-2023)
 ----------------------------